fix(i18n): repair seed generation and isolate notification flows
