+++ .fallowrc.json
@@ -38,6 +38,7 @@
     "apps/web/src/index.css",
     "apps/web/src/components/ui/**",
     "scripts/seed.ts",
+    "scripts/seed-main.ts",
     "packages/i18n/src/generated/**",
     "packages/i18n/.generated/**"
   ],
